Added short output, updates module, SSH password, defaults
The short output will show a briefer version of normal output for
plain. JSON output is unaffected. Currently only the needs_upgrades
module has short output.

Added needs_upgrades module. The module will return all packages
that are available for upgrade. Works with APT, DNF, and YUM.

An SSH password can be used instead of a private key.

A few configuration settings are set to defaults if they're not set
by the user.