Created Go script to generate module struct types
The script uses a special comment in module files that tells it
if a module outputs a JSON object or array and generates the
appropiate struct type for the response. It will also generate
print functions. Currently only the long print options are available.

Going forward, src/utils/moduleTypes.go should never be hand
edited. If a fix is needed, edit the generation script or the
gen comment in the module.