Fixed disk_part mod, renamed ip_addresses, don't die on down host
The disk_partitions module still had a percent sign in one of the
key names. That has sense been removed. The used percentage is also
an integer now instead of a string. The generator script has been
modified to allow percent signs as a suffix.

The ip_addresses module has been renamed to network_interfaces.
Eventually that module will be expanded to include MAC addresses,
gateways, and subnet masks. Obviously much more than just ip
addresses.

A run won't die if a host in unavailable. The timeout is configurable
using the SSH.Timeout key. It must be valid Go Time.Duration string.
The default is 20 seconds. If a host is unavailable for any reason,
it's marked as disabled for the remander of the run.