Fixed network_interfaces module for newer Linux systems
Apparently "ip" is the new, recommended command for getting
network interface information. Ifconfig is so last year. Plus its
syntax is inconsistant between versions its a bit tricky to parse
correctly. The module will use "ip" if it's available and fallback
to "ifconfig" if it's not. Worst case would be neither is available
and it outputs an empty array.