Use gzip compression for binary data
The generated source file was getting too big. Using compression,
the file size is reduced by almost half. There's not much affect
on the binary itself. Both before and after the change the binary
was 5.4MB. But then again, it's only a change of a few KB so much
compared to MB. Still, it's something.