Added download instructions, ensure "unix" paths in static assets
When running `make generate` from a Windows machine, the asset
paths had Windows directory separator "\". This caused the entire
file to be re-generated and would have broken if built on Windows
targeting Linux. All assets are set to a unix like path for consistency
and because the unix directory separator "/" isn't an escape character.