Adds host specific connection options per usi-lfkeitel/saturn#2
I have not tested this just yet, but it compiles. It appends a
password and or private key authentication method to the sshClientConfig
object given to the host when connecting. The host stores a copy of the
global config possibly with its own changes. The username, password,
and private key file can be overriden for each host.